BUGFIX: Use rails convention to underscore collection class names
This would mean a class such as `ActionItem` when evaluated in a partial would show up as `actionitem` instead of the expected `action_item`. Users of the library could override this with the `as:` option.

Added extensive tests to ensure broadcast and `futurize` generator properly generate
Note this is a breaking change for users, but is correct in following rails conventions.